fix(memberships): guard None row in add_membership, add POST remaining regression test
Code review on 21b6770 flagged a TypeError risk if the post-upsert re-read
races with a concurrent delete, and a missing POST-side counterpart to
test_patch_remaining_is_ignored. Also widens MembershipReadSerializer.remaining
to FloatField for half-lesson correctness (currently dead code, not wired
into any view).

+++ journal_django/apps/memberships/serializers.py
@@ -27,7 +27,7 @@ class MembershipReadSerializer(serializers.Serializer):
     group_id = serializers.IntegerField()
     student_id = serializers.IntegerField()
     lessons_done = serializers.IntegerField()
-    remaining = serializers.IntegerField()
+    remaining = serializers.FloatField()
     start_date = DateStringField(allow_null=True)
     sheet_row = serializers.IntegerField(allow_null=True)
     active = serializers.BooleanField()
